refactor: FirestoreServerVideoData型を循環参照解消のために移動
- packages/shared-types/src/types/firestore/ディレクトリを新規作成
- FirestoreServerVideoDataとその関連型をvideo.tsからtypes/firestore/video.tsに移動
- 循環参照を防ぐため、エンティティとPlain Object間の依存関係を整理
- 全てのインポートパスを新しい場所に更新
- ENTITY_IMPLEMENTATION_GUIDELINES.mdにFirestore型配置方針を追加

また、youtube-service.tsにデバッグログを追加してエラーハンドリングを改善

Copilotレビューの指摘事項への対応:
- 循環参照の可能性を解消（最優先事項）
- エラーハンドリングの追加
