+++ packages/shared-types/src/entities/video.ts
@@ -7,6 +7,11 @@
  */
 
 import type { VideoComputedProperties, VideoPlainObject } from "../plain-objects/video-plain";
+import type {
+	FirestoreServerVideoData,
+	LiveBroadcastContent,
+	VideoType,
+} from "../types/firestore/video";
 import { Channel } from "../value-objects/video/channel";
 import {
 	ContentDetails,
@@ -49,113 +54,21 @@ export interface AudioButtonInfo {
 	hasButtons: boolean;
 }
 
-/**
- * Legacy types for backward compatibility
- */
-export type LiveBroadcastContent = "none" | "live" | "upcoming";
-export type VideoType = "normal" | "archived" | "premiere";
-
 // Re-export Plain Object types from separate file
 export type { VideoComputedProperties, VideoPlainObject } from "../plain-objects/video-plain";
+
+// Re-export Firestore types for convenience
+export type {
+	FirestoreServerVideoData,
+	FirestoreVideoData,
+	LiveBroadcastContent,
+	VideoType,
+} from "../types/firestore/video";
 
 /**
  * Frontend video data type - Plain object for Server/Client Component boundary
  */
 export type FrontendVideoData = VideoPlainObject;
-
-/**
- * Firestore server video data type
- */
-export interface FirestoreServerVideoData {
-	id?: string;
-	videoId: string;
-	title: string;
-	description: string;
-	channelId: string;
-	channelTitle: string;
-	publishedAt: unknown; // Firestore Timestamp
-	thumbnailUrl: string;
-	lastFetchedAt: unknown; // Firestore Timestamp
-	videoType?: VideoType;
-	liveBroadcastContent?: LiveBroadcastContent;
-
-	// Audio button info
-	audioButtonCount?: number;
-	hasAudioButtons?: boolean;
-
-	// Content details
-	duration?: string;
-	categoryId?: string;
-	tags?: string[];
-	playlistTags?: string[];
-	userTags?: string[];
-
-	// Statistics
-	statistics?: {
-		viewCount?: number;
-		likeCount?: number;
-		dislikeCount?: number;
-		favoriteCount?: number;
-		commentCount?: number;
-	};
-
-	// Live streaming details
-	liveStreamingDetails?: {
-		scheduledStartTime?: unknown; // Firestore Timestamp
-		scheduledEndTime?: unknown; // Firestore Timestamp
-		actualStartTime?: unknown; // Firestore Timestamp
-		actualEndTime?: unknown; // Firestore Timestamp
-		concurrentViewers?: number;
-	};
-
-	// Video quality details
-	dimension?: string;
-	definition?: string;
-	caption?: string | boolean; // Can be either string or boolean for compatibility
-	licensedContent?: boolean;
-
-	// Video status
-	status?: {
-		privacyStatus?: string;
-		uploadStatus?: string;
-		embeddable?: boolean;
-		publicStatsViewable?: boolean;
-		madeForKids?: boolean;
-		selfDeclaredMadeForKids?: boolean;
-		commentStatus?: string;
-	};
-
-	// Player embed
-	player?: {
-		embedHtml?: string;
-		embedWidth?: number;
-		embedHeight?: number;
-	};
-
-	// Topic details
-	topicDetails?: {
-		topicCategories?: string[];
-	};
-
-	// Recording details
-	recordingDetails?: {
-		location?: string;
-		locationDescription?: string;
-		recordingDate?: string;
-	};
-
-	// Region restriction
-	regionRestriction?: {
-		allowed?: string[];
-		blocked?: string[];
-	};
-
-	// Content rating
-	contentRating?: Record<string, unknown>;
-}
-
-// Alias for backward compatibility
-export type FirestoreVideoData = FirestoreServerVideoData;
 
 /**
  * Video list result type
