+++ packages/shared-types/src/types/firestore/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Firestore type definitions
+ *
+ * Export all Firestore-related type definitions from this module
+ */
+
+export * from "./video";
